+++ core/cmake/BareosCheckFunctions.cmake
@@ -1,6 +1,6 @@
 #   BAREOS® - Backup Archiving REcovery Open Sourced
 #
-#   Copyright (C) 2017-2020 Bareos GmbH & Co. KG
+#   Copyright (C) 2017-2021 Bareos GmbH & Co. KG
 #
 #   This program is Free Software; you can redistribute it and/or
 #   modify it under the terms of version three of the GNU Affero General Public
@@ -87,3 +87,5 @@ check_function_exists(unlinkat HAVE_UNLINKAT)
 check_function_exists(utimes HAVE_UTIMES)
 
 check_function_exists(glfs_readdirplus HAVE_GLFS_READDIRPLUS)
+
+check_function_exists(chflags HAVE_CHFLAGS)